pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hf_ehr"
version = "0.1.3"
authors = [
    { name = "Michael Wornow" }
]
description = "Code for Context Clues paper"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "polars==1.1",
    "transformers==4.48.2",
    "trl==0.10.1",
    "datasets==2.14.5",
    "pandas==2.2",
    "tensorboard==2.14.1",
    "lightning==2.2.1",
    "numpy==1.26.0",
    "pytorch-nlp==0.5.0",
    "wandb==0.15.11",
    "loguru==0.7.2",
    "omegaconf==2.3.0",
    "jaxtyping==0.2.22",
    "mlflow==2.11.1",
    "seaborn==0.13.2",
    "matplotlib==3.9.2",
    "hydra-core==1.3.2",
    "lightgbm==3.3.5",
    "scikit-learn==1.2.2",
    "scipy==1.10.1",
    "tqdm==4.65.0",
    "tabulate==0.9.0",
    "huggingface_hub==0.26.5",
    "meds==0.3.3",
    "meds_reader==0.1.9",
    "meds_etl==0.3.8",
    "femr==0.1.16",
    "pyyaml==6.0.1",
    "torch"
]

[project.urls]
Homepage = "https://github.com/som-shahlab/long_context_clues/"

[tool.setuptools.packages.find]
include = ["hf_ehr*"]